refactor for better code reusability
