+++ .gitignore
@@ -131,3 +131,4 @@ dmypy.json
 
 #Models
 saved_agent/*
+outputs/
