+++ src/agents/mc_agent.py
@@ -41,32 +41,8 @@ class MonteCarloNN(BaseAgent):
         if optimizer:
             self.optimizer.load_state_dict(optimizer)
         self.losses = []
-        self.random_count = 0
-        self.greedy_count = 0
-        self.action_count = [0, 0, 0, 0]
         self.mode = mode
 
-    def select_action(self, observation):
-        self.prev_state = self._get_state_from_obs(observation)
-        max_value = -100
-        max_index = 0
-        with torch.no_grad():
-            p = np.random.random()
-            if p > self.epsilon:
-                self.greedy_count += 1
-                for action_index, move in self._action_to_direction.items():
-                    state_tensor = self.prev_state.clone()
-                    state_tensor[:2, :] += torch.from_numpy(move).reshape((-1, 1))
-                    value = self.value_function(state_tensor.T)
-                    if value.item() > max_value:
-                        max_value = value.item()
-                        max_index = action_index
-            else:
-                max_index = np.random.randint(0, 3 + 1)
-                self.random_count += 1
-
-        self.action_count[max_index] += 1
-        return max_index
 
     def reset(self):
         self.state_sequence = []
@@ -80,30 +56,11 @@ class MonteCarloNN(BaseAgent):
             loss = self._train(Gs)
             self.losses.append(loss)
 
-    def save(self, filename):
-        torch.save(
-            {
-                "value_function": self.value_function.state_dict(),
-                "optimizer_state": self.optimizer.state_dict(),
-                "loss": self.loss_function.state_dict(),
-                "losses": self.losses,
-            },
-            filename,
-        )
 
     def eval(self):
         self.epsilon /= 100
         self.mode = "testing"
-        self.action_count = [0, 0, 0, 0]
-        self.greedy_count = 0
-        self.random_count = 0
 
-    def _get_state_from_obs(self, observation):
-        head = observation["agent"]
-        target = observation["target"]
-        body = observation["body"]
-        state_array = np.concatenate((head, target, body[0]), dtype=np.float32)
-        return torch.from_numpy(state_array).reshape((-1, 1))
 
     def _compute_Gs(self):
         G = self.state_sequence[-1][-1]
