+++ src/agents/td_agent.py
@@ -45,25 +45,6 @@ class TDLambdaNN(BaseAgent):
         self.mode = mode
 
 
-    def select_action(self, observation):
-        self.prev_state = self._get_state_from_obs(observation)
-        max_value = -100
-        max_index = 0
-        with torch.no_grad():
-            p = np.random.random()
-            if p > self.epsilon:
-                for action_index, move in self._action_to_direction.items():
-                    state_tensor = self.prev_state.clone()
-                    state_tensor[:2, :] += torch.from_numpy(move).reshape((-1, 1))
-                    value = self.value_function(state_tensor.T)
-                    if value.item() > max_value:
-                        max_value = value.item()
-                        max_index = action_index
-            else:
-                max_index = np.random.randint(0, 3 + 1)
-
-        return max_index
-
     def update(self, reward, observation, action, terminated):
         #we voluntiraly don't zero grad the optimizer to consider eligibility traces
         for parameter in self.value_function.parameters():
@@ -81,26 +62,7 @@ class TDLambdaNN(BaseAgent):
     def reset(self):
         return None
 
-    def save(self,filename):
-        #TODO make this a function in the base agent
-        torch.save(
-            {
-                "value_function": self.value_function.state_dict(),
-                "optimizer_state": self.optimizer.state_dict(),
-                "loss": self.loss_function.state_dict(),
-            },
-        filename,
-    )
 
     def eval(self):
         self.epsilon /= 100
         self.mode = "testing"
-
-
-    def _get_state_from_obs(self,observation):
-        #TODO make this a function in the base agent
-        head = observation["agent"]
-        target = observation["target"]
-        body = observation["body"]
-        state_array = np.concatenate((head, target, body[0]), dtype=np.float32)
-        return torch.from_numpy(state_array).reshape((-1, 1))
