+++ src/util.py
@@ -1,6 +1,7 @@
 import torch
 from controller import agent_mode
 import envs
+import numpy as np
 
 def initialize_new_agent(agent, **kwargs):
     return agent(**kwargs)
@@ -15,11 +16,13 @@ def load_agent(agent, filename, **kwargs):
 def train_and_save(agent, n_episodes, max_step, filename):
     env = envs.SnakeEnv(size=10)
     _,_ = env.reset()
-    agent_mode(env=env, n_episodes=n_episodes, agent=agent, max_step=max_step)
+    rewards = agent_mode(env=env, n_episodes=n_episodes, agent=agent, max_step=max_step)
+    with open('outputs/output.txt', 'ab') as f:
+        f.write(b'\n')
+        np.savetxt(f, rewards)
     agent.save(filename)
 
 def test(agent, n_episodes, max_step):
     env = envs.SnakeEnv(render_mode='human', size=10)
     agent.eval()
     agent_mode(env=env, n_episodes=n_episodes, agent=agent, max_step=max_step, mode='testing')
-
